passing down updateEventAttendance as props
